Refactor: Replace macros with enum class trig_fn

+++ common/encoder.cpp
@@ -739,10 +739,10 @@ static void setup_save_info(FRACTAL_INFO *save_info)
         save_info->bailoutold = 0;
 
     save_info->calctime = calctime;
-    save_info->trigndx[0] = trigndx[0];
-    save_info->trigndx[1] = trigndx[1];
-    save_info->trigndx[2] = trigndx[2];
-    save_info->trigndx[3] = trigndx[3];
+    save_info->trigndx[0] = static_cast<BYTE>(trigndx[0]);
+    save_info->trigndx[1] = static_cast<BYTE>(trigndx[1]);
+    save_info->trigndx[2] = static_cast<BYTE>(trigndx[2]);
+    save_info->trigndx[3] = static_cast<BYTE>(trigndx[3]);
     save_info->finattract = (short) (finattract ? 1 : 0);
     save_info->initorbit[0] = initorbit.x;
     save_info->initorbit[1] = initorbit.y;
